# specs/001-autonomous-education-platform/contracts/education-api.openapi.yaml
openapi: 3.0.3
info:
  title: Kidpen Education Platform API
  version: 0.1.0
  description: Initial contract for the autonomous education platform vertical slices.
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /education/tutor/messages:
    post:
      summary: Send a tutor message and start or continue a production agent run
      operationId: sendTutorMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TutorMessageRequest'
      responses:
        '202':
          description: Tutor run accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorRunAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /education/tutor/threads/{threadId}:
    get:
      summary: Get tutor thread metadata and permitted messages
      operationId: getTutorThread
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Thread returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TutorThread'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Not authorized for this resource or action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        reasonCode:
          type: string
...
    TutorMessageRequest:
      type: object
      required: [message]
      properties:
        threadId: { type: string }
        subjectKey: { type: string }
        knowledgeComponentId: { type: string }
        assignmentId: { type: string }
        message: { type: string }
        clientContext:
          type: object
          additionalProperties: true
    TutorRunAccepted:
      type: object
      required: [threadId, runId, streamUrl, policyMode]
      properties:
        threadId: { type: string }
        runId: { type: string }
        streamUrl: { type: string }
        policyMode: { type: string, enum: [socratic, homework_guidance, safety, fallback] }
    TutorThreadSummary:
      type: object
      required: [id, title, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        subjectKey: { type: string }
        updatedAt: { type: string, format: date-time }
    TutorThread:
      allOf:
        - $ref: '#/components/schemas/TutorThreadSummary'
        - type: object
          properties:
            messages:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  role: { type: string }
                  content: { type: string }
                  createdAt: { type: string, format: date-time }
...